% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bilinear_interpolation}
\alias{bilinear_interpolation}
\title{bilinear_interpolation}
\usage{
bilinear_interpolation(
  coord,
  from_corner_coord,
  to_corner_coord,
  ordered_corner = F
)
}
\description{
bilinear_interpolation
}
