% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{XY_computation}
\alias{XY_computation}
\title{XY_computation}
\usage{
XY_computation(
  Fieldplot_BDD_full,
  raw_files_path,
  subplot,
  plot_name,
  files_Summary,
  directory
)
}
\description{
XY_computation
}
